move namedtempfile to builder
